+++ src/binary/binary.c
@@ -21,6 +21,7 @@
 #include "utils/palloc.h"
 
 #define CHC_IMPLEMENTATION
+#define CHC_NO_ASYNC
 #include "clickhouse.h"
 #include "clickhouse-compression.h"
 #include "clickhouse-posix-io.h"