+++ src/deparse.c
@@ -1789,7 +1789,7 @@ deparseRelation(StringInfo buf, Relation rel)
 	ForeignServer *server = chfdw_get_foreign_server(rel);
 	ListCell   *lc;
 
-	chfdw_extract_options(server->options, NULL, NULL, NULL, &dbname, NULL, NULL);
+	chfdw_extract_options(server->options, NULL, NULL, NULL, &dbname, NULL, NULL, NULL);
 
 	/* obtain additional catalog information. */
 	table = GetForeignTable(RelationGetRelid(rel));